Use ellipsis for text overflow
I had a dropdown menu item that had text selections that broke onto 2 lines. I updated my local CSS to use ellipsis instead of breaking into a 2nd line. This meant changing how the carate was positioned in the CSS. It no longer floats but is positioned absolutely. 

Sharing this in case it's helpful.